Remove unused requires of Google Closure classes (#15)
Everything in the Google Closure Library is available without any explicit require or import as long as the names are fully qualified with `goog....`. That is the form we use here, so no require or import is necessary.

Classes like NumberFormat and DateTimeFormat could be made available without the `goog.` prefix or be renamed with an `import`, but qualifying them fully works as well.